fixed frontend hooks
