+++ web/src/app/chat/chat_search/ChatSearchGroup.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { ChatSearchItem } from "./ChatSearchItem";
-import { ChatSessionSummary } from "./api/interfaces";
+import { ChatSessionSummary } from "./interfaces";
 
 interface ChatSearchGroupProps {
   title: string;
