--- web/src/app/chat/chat_search/components/index.ts
@@ -1,1 +1,0 @@
-export * from "./SearchInput";
