--- web/src/app/chat/chat_search/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from "./useIntersectionObserver";
